refactor: extract /agents, /skills, /backends CRUD into internal/server/fleet
Phase 3b of the webhook split (issue #250). Moves the agent, skill, and
backend HTTP CRUD surface plus the methods exposed for the MCP
AgentWriter / SkillWriter / BackendWriter interfaces into a new
dedicated package.

Architecture:
  - internal/server/fleet/Handler holds the new home for handlers,
    server methods, wire types (storeAgentJSON, agentPatch,
    storeSkillJSON, skillPatch, storeBackendJSON, backendPatchJSON,
    localBackendRequest), and the MCP-adapter conversions.
  - The previous internal/webhook/mcp_adapters.go is gone; its three
    adapter methods now live on fleet.Handler directly, converting
    mcp.AgentPatch / SkillPatch / BackendPatch in-package.
  - webhook/Server constructs the fleet handler in WithStore and
    exposes it via Fleet() so cmd/agents wires it as
    AgentWrite/SkillWrite/BackendWrite for MCP tools.
  - GET /agents (the read-only fleet snapshot view) and
    /agents/orphans/status stay in webhook for now; their /status
    interplay will move with the orphans + fleet snapshot extraction
    in a follow-up PR.

webhook/crud.go shrinks ~840 lines. No behaviour change: the same
routes respond on the same paths, the same wire shapes are used, the
same locking discipline applies (one storeMu, one reload epoch per
write through the WriteCoordinator). Tests in internal/webhook still
exercise the full router via buildHandler so they cover the new
package without modification; a small crud_wire_test.go declares
local mirrors of the response shapes so test assertions don't depend
on the new package's unexported types.
