+++ cmd/agents/main.go
@@ -170,9 +170,9 @@ func run() error {
 		Memory:        &sqliteMcpReader{db: db},
 		ConfigBytes:   srv,
 		ConfigImport:  srv,
-		AgentWrite:    srv,
-		SkillWrite:    srv,
-		BackendWrite:  srv,
+		AgentWrite:    srv.Fleet(),
+		SkillWrite:    srv.Fleet(),
+		BackendWrite:  srv.Fleet(),
 		RepoWrite:     srv.Repos(),
 		BindingWrite:  srv.Repos(),
 		Logger:        logger,
